Move encoder tests to separate file
