+++ tests/test.js
@@ -3,7 +3,7 @@ const path = require('path');
 const { test } = require('uvu');
 const assert = require('uvu/assert');
 
-const { NbsDecoder, NbsEncoder } = require('..');
+const { NbsDecoder } = require('../nbsdecoder');
 
 /** Convert the given timestamp object to a BigInt of nanoseconds */
 function tsToBigInt(ts) {
@@ -436,154 +436,6 @@ test('NbsDecoder.getPackets() returns the last packet of each type when given a 
     subtype: 0,
     payload: Buffer.from('pong.699', 'utf8'),
   });
-});
-
-test('NbsEncoder constructor throws for invalid arguments', () => {
-  assert.throws(
-    () => {
-      new NbsEncoder();
-    },
-    /missing argument `path`: provide a path to write to/,
-    'NbsEncoder() constructor throws without path argument'
-  );
-
-  assert.throws(
-    () => {
-      new NbsEncoder(1);
-    },
-    /invalid argument `path`: expected string/,
-    'NbsEncoder() constructor throws with incorrect type for path argument'
-  );
-});
-
-test('NbsEncoder write throws for invalid arguments', () => {
-  const writePath = path.join(samplesDir, 'writeTest.nbs');
-  const encoder = new NbsEncoder(writePath);
-
-  assert.throws(
-    () => {
-      encoder.write();
-    },
-    /missing argument `packet`: provide a packet to write to the file/,
-    'NbsEncoder write() throws without packet argument'
-  );
-
-  assert.throws(
-    () => {
-      encoder.write({});
-    },
-    /invalid type for argument `packet`: expected object with `timestamp` key/,
-    'NbsEncoder write() throws without timestamp in packet'
-  );
-
-  assert.throws(
-    () => {
-      encoder.write({
-        timestamp: { seconds: 1897, nanos: 0 },
-      });
-    },
-    /invalid type for argument `packet`: expected object with `type` key/,
-    'NbsEncoder write() throws without type in packet'
-  );
-
-  assert.throws(
-    () => {
-      encoder.write({
-        timestamp: { seconds: 1897, nanos: 0 },
-        type: pingType,
-      });
-    },
-    /invalid type for argument `packet`: expected object with `subtype` key/,
-    'NbsEncoder write() throws without subtype in packet'
-  );
-
-  assert.throws(
-    () => {
-      encoder.write({
-        timestamp: { seconds: 1897, nanos: 0 },
-        type: pingType,
-        subtype: 0,
-      });
-    },
-    /invalid type for argument `packet`: expected object with `payload` key/,
-    'NbsEncoder write() throws without payload in packet'
-  );
-
-  assert.throws(
-    () => {
-      encoder.write({
-        timestamp: 'string',
-        type: pingType,
-        subtype: 0,
-        payload: Buffer.from('ping.699', 'utf8'),
-      });
-    },
-    /invalid type for argument `packet`: error in `timestamp`: expected positive number or BigInt or timestamp object/,
-    'NbsEncoder write() throws with incorrect type for timestamp'
-  );
-
-  assert.throws(
-    () => {
-      encoder.write({
-        timestamp: { wrongKey: 1 },
-        type: pingType,
-        subtype: 0,
-        payload: Buffer.from('ping.699', 'utf8'),
-      });
-    },
-    /invalid type for argument `packet`: error in `timestamp`: expected object with `seconds` and `nanos` keys/,
-    'NbsEncoder write() throws for incorrect keys in timestamp'
-  );
-
-  assert.throws(
-    () => {
-      encoder.write({
-        timestamp: { seconds: 'string', nanos: 'string' },
-        type: pingType,
-        subtype: 0,
-        payload: Buffer.from('ping.699', 'utf8'),
-      });
-    },
-    /invalid type for argument `packet`: error in `timestamp`: `seconds` and `nanos` must be numbers/,
-    'NbsEncoder write() throws with incorrect types for timestamp properties'
-  );
-
-  // Delete file written to
-  fs.unlinkSync(writePath);
-  fs.unlinkSync(writePath + '.idx');
-});
-
-test('Packets written by NbsEncoder can be read by NbsDecoder', () => {
-  const writePath = path.join(samplesDir, 'writeTest.nbs');
-  const encoder = new NbsEncoder(writePath);
-
-  const pingPacket = {
-    timestamp: { seconds: 1897, nanos: 0 },
-    type: pingType,
-    subtype: 0,
-    payload: Buffer.from('ping.699', 'utf8'),
-  };
-  const pongPacket = {
-    timestamp: { seconds: 1898, nanos: 0 },
-    type: pongType,
-    subtype: 0,
-    payload: Buffer.from('pong.699', 'utf8'),
-  };
-
-  encoder.write(pingPacket);
-  encoder.write(pongPacket);
-  encoder.close();
-
-  const decoder = new NbsDecoder([writePath]);
-  const packets = decoder.getPackets({ seconds: 1899, nanos: 0 });
-
-  assert.equal(packets.length, 2);
-  assert.equal(packets[0], pingPacket);
-  assert.equal(packets[1], pongPacket);
-
-  // Delete file written to
-  fs.unlinkSync(writePath);
-  fs.unlinkSync(writePath + '.idx');
 });
 
 test.run();
